+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/SettingsViewModel.kt
@@ -5,13 +5,11 @@ import androidx.datastore.core.DataStore
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
-import com.flipperdevices.core.di.ApplicationParams
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.error
 import com.flipperdevices.core.preference.pb.SelectedTheme
 import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.share.ShareHelper
-import com.flipperdevices.selfupdater.api.SelfUpdaterApi
 import com.flipperdevices.settings.impl.R
 import com.flipperdevices.settings.impl.model.ExportState
 import com.flipperdevices.shake2report.api.Shake2ReportApi
@@ -25,14 +23,11 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import tangle.viewmodel.VMInject
 
-@Suppress("LongParameterList", "TooManyFunctions")
 class SettingsViewModel @VMInject constructor(
     private val dataStoreSettings: DataStore<Settings>,
-    private val applicationParams: ApplicationParams,
     private val exportKeysHelper: ExportKeysHelper,
     private val shake2ReportFeatureEntry: Shake2ReportFeatureEntry,
     private val shake2ReportApi: Shake2ReportApi,
-    private val selfUpdaterApi: SelfUpdaterApi
 ) : ViewModel(), LogTagProvider {
     override val TAG = "SettingsViewModel"
 
@@ -116,7 +111,4 @@ class SettingsViewModel @VMInject constructor(
             }
         }
     }
-
-    fun versionApp() = applicationParams.version
-    fun sourceInstall() = selfUpdaterApi.getInstallSourceName()
 }