+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 1.6.5 - In progress
 
+- [Feature] Fap Catalog save sort
+- [Feature] Check Self Update App in Optiond (only for github)
 - [FIX] Use by default dark theme in Wear OS
 - [FIX] Use default splashscreen in Wear OS
 - [FIX] Fix compose layout in Wear OS
@@ -14,6 +16,7 @@
 - [Feature] Add fap manifest caching
 - [Feature] Not send empty report in faphub
 - [Feature] Add hide app button
+- [Feature] Check Self Update App in Optiond (only for github)
 - [CI] Add apk artifacts to Pull Requests
 - [CI] Fix failed build in merge_group
 - [CI] Bump deps except Sentry-gradle